push

+++ frontend/src/components/WorkoutDetails.js
@@ -2,14 +2,24 @@ import { useWorkoutsContext } from "../hooks/useWorkoutsContext"
 
 //date-fns
 import formatDistanceToNow from "date-fns/formatDistanceToNow"
+import { useAuthContext } from "../hooks/useAuthContext"
 
 
 const WorkourDetails = ({workout}) => {
   const { dispatch } = useWorkoutsContext()
+  const { user } = useAuthContext()
 
   const handleClick = async () => {
+
+    if(!user){
+      return
+    }
+
     const response = await fetch("api/workouts/" + workout._id, {
       method: "DELETE",
+      headers: {
+        "Authorization": `Bearer ${user.token}`    // backticks for template string to have variable inside
+      }
     })
 
     const json = await response.json()
